requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.4
Flask==3.1.0
Flask-Cors==5.0.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
numpy>=1.24,<2.0
packaging==24.2
pandas>=2.2.0
requests==2.25.1
scikit-learn>=1.4,<1.6
Werkzeug==3.1.3
torch>=2.2.0
torchvision>=0.17
timm>=0.9.12
Pillow>=10.0.0
